Valida se usuário que quer modificar item é dono do item, adiciona categorias para itens, adiciona sistema de cache com redis para itens mais acessados

+++ src/menu-items/menu-items.service.ts
@@ -1,15 +1,29 @@
-import { ConflictException, Injectable } from "@nestjs/common";
+import { Restaurant } from "src/restaurants/restaurant.entity";
+import { User } from "src/users/user.entity";
+import { Repository } from "typeorm";
+
+import {
+  ConflictException,
+  Inject,
+  Injectable,
+  UnauthorizedException
+} from "@nestjs/common";
+import { InjectRepository } from "@nestjs/typeorm";
+
 import { CreateMenuItemDto } from "./dto/create-menu-item.dto";
 import { UpdateMenuItemDto } from "./dto/update-menu-item.dto";
-import { InjectRepository } from "@nestjs/typeorm";
 import { MenuItem } from "./menu-item.entity";
-import { Repository } from "typeorm";
+import { MenuItemsCacheService } from "src/menu-items/menu-items-cache.service";
 
 @Injectable()
 export class MenuItemsService {
   constructor(
     @InjectRepository(MenuItem)
-    private menuItemRepository: Repository<MenuItem>
+    private menuItemRepository: Repository<MenuItem>,
+    @InjectRepository(Restaurant)
+    private restaurantRepository: Repository<Restaurant>,
+    @Inject(MenuItemsCacheService)
+    private menuItemsCacheService: MenuItemsCacheService
   ) {}
 
   async create(createMenuItemDto: CreateMenuItemDto): Promise<MenuItem> {
@@ -28,13 +42,13 @@ export class MenuItemsService {
   async findAll(
     page: number = 1,
     limit: number = 10
-  ): Promise<{ data: MenuItem[]; totalPages: number }> {
+  ): Promise<{ data: MenuItem[]; totalPages: number; total: number }> {
     const [menuItems, total] = await this.menuItemRepository.findAndCount({
       skip: (page - 1) * limit,
       take: limit
     });
     const totalPages = Math.ceil(total / limit);
-    return { data: menuItems, totalPages };
+    return { data: menuItems, totalPages, total };
   }
 
   async findOne(id: string): Promise<MenuItem> {
@@ -46,32 +60,68 @@ export class MenuItemsService {
       throw new ConflictException("Item de cardápio não encontrado");
     }
 
+    menuItem.accessCount += 1;
+    await this.menuItemRepository.save(menuItem);
+    await this.updatePopularCache();
+
     return menuItem;
+  }
+
+  async updatePopularCache() {
+    const popularItems = await this.menuItemRepository.find({
+      order: { accessCount: "DESC" },
+      take: 5
+    });
+
+    await this.menuItemsCacheService.setPopularMenuItems(popularItems);
   }
 
   async update(
     id: string,
+    user: User,
     updateMenuItemDto: UpdateMenuItemDto
   ): Promise<UpdateMenuItemDto> {
-    const existingItem = await this.menuItemRepository.findOneBy({
-      id: updateMenuItemDto.id
-    });
+    const existingItem = await this.menuItemRepository.findOneBy({ id });
     if (!existingItem) {
       throw new ConflictException("Item de cardápio não existe");
     }
 
-    await this.menuItemRepository.update(id, updateMenuItemDto);
+    const restaurant = await this.restaurantRepository.findOne({
+      where: { id: existingItem.restaurantId }
+    });
 
-    return updateMenuItemDto;
+    const isUserOwner = restaurant.userId === user.id;
+
+    if (isUserOwner) {
+      Object.assign(existingItem, updateMenuItemDto);
+      await this.menuItemRepository.save(existingItem);
+      return updateMenuItemDto;
+    } else {
+      throw new UnauthorizedException(
+        "Item do menu só pode ser atualizado pelo usuário responsável"
+      );
+    }
   }
 
-  async remove(id: string): Promise<void> {
+  async remove(id: string, user: User): Promise<void> {
     const item = await this.findOne(id);
 
     if (!item) {
       throw new ConflictException("Item de cardápio não encontrado");
     }
 
-    await this.menuItemRepository.remove(item);
+    const restaurant = await this.restaurantRepository.findOne({
+      where: { id: item.restaurantId }
+    });
+
+    const isUserOwner = restaurant.userId === user.id;
+
+    if (isUserOwner) {
+      await this.menuItemRepository.remove(item);
+    } else {
+      throw new UnauthorizedException(
+        "Item do menu só pode ser removido pelo usuário responsável"
+      );
+    }
   }
 }
